Add NFBackButtonListener
This is a replacement for the builtin BackButtonListener, which solves
the following problem:

If a BackButtonListener exists in the widget tree and the tree gets
rebuilt (e.g by rotating the phone or showing the software keyboard),
it re-registers its listener. This re-registration causes the listener
to take priority over other listeners, which don't get re-registered
when the widget tree gets rebuilt. An example of such a listener is a
manually registered listener.