Listen for back button presses when showing a dialog
BackButtonListeners take precedence over the default behaviour, which is
generating a Navigator.pop call. Therefore the listeners stole the back
button event before the dialog was notified.

Register a back button listener that takes priority and pops the dialog
while the dialog is shown. We can't use a BackButtonListener here
because we are not building a widget tree. We also can't use a
BackButtonListener as content of the dialog, because it gets inserted
very close to the root of the widget tree, which causes other
BackButtonListeners lower in the widget tree to have priority.